[Rewards] Move top-level page components into appropriate folders
