+++ components/brave_rewards/resources/rewards_page/components/app.tsx
@@ -16,15 +16,15 @@ import { HomeView } from './home/home_view'
 import { ExploreView } from './explore/explore_view'
 import { Onboarding } from './onboarding/onboarding'
 import { OnboardingSuccessModal } from './onboarding/onboarding_success_modal'
-import { ConnectAccount } from './connect_account'
+import { ConnectAccount } from './connect/connect_account'
 import { UnsupportedRegionView } from './unsupported_region_view'
-import { AuthorizationModal } from './authorization_modal'
-import { CaptchaModal } from './captcha_modal'
+import { AuthorizationModal } from './connect/authorization_modal'
+import { CaptchaModal } from './modals/captcha_modal'
 import { ContributeModal } from './contribute/contribute_modal'
-import { ResetModal } from './reset_modal'
-import { TosUpdateModal } from './tos_update_modal'
-import { NotificationModal } from './notification_modal'
-import { SelfCustodyInviteModal } from './self_custody_invite_modal'
+import { ResetModal } from './modals/reset_modal'
+import { TosUpdateModal } from './modals/tos_update_modal'
+import { NotificationModal } from './modals/notification_modal'
+import { SelfCustodyInviteModal } from './modals/self_custody_invite_modal'
 import { useShouldShowSelfCustodyInvite } from '../lib/self_custody_invite'
 import { useConnectAccountRouter } from '../lib/connect_account_router'
 import * as routes from '../lib/app_routes'
