+++ components/brave_rewards/resources/rewards_page/components/common/modal.style.ts
@@ -4,7 +4,7 @@
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 import { color, font } from '@brave/leo/tokens/css/variables'
-import { scoped } from '../lib/scoped_css'
+import { scoped } from '../../lib/scoped_css'
 
 export const modalStyle = scoped.css`
   & {
