+++ components/brave_rewards/resources/rewards_page/components/common/modal.tsx
