+++ components/brave_rewards/resources/rewards_page/components/common/text_skeleton.tsx
