+++ components/brave_rewards/resources/rewards_page/components/connect/authorization_modal.style.ts
@@ -4,25 +4,26 @@
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 import { color } from '@brave/leo/tokens/css/variables'
-import { scoped } from '../lib/scoped_css'
+import { scoped } from '../../lib/scoped_css'
 
 export const style = scoped.css`
   & {
-    overflow: auto;
-    max-width: 600px;
-    color: ${color.text.secondary};
-
-    a {
-      font-weight: 600;
-      text-decoration: none;
-    }
-
     @container style(--is-wide-view) {
       max-width: 600px;
     }
   }
 
-  .message-icon {
+  .processing {
+    --leo-progressring-size: 38px;
+
+    padding: 32px 0 16px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 32px;
+  }
+
+  .status-icon {
     --leo-icon-size: 40px;
 
     height: 88px;
@@ -34,8 +35,14 @@ export const style = scoped.css`
     color: ${color.systemfeedback.warningIcon};
   }
 
-  .text {
-    padding-bottom: 16px;
+  h3 {
+    margin: 32px 0 16px;
+  }
+
+  .error-text {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
   }
 `
