+++ components/brave_rewards/resources/rewards_page/components/contribute/contribute_modal.tsx
@@ -8,7 +8,7 @@ import * as React from 'react'
 import { AppModelContext, useAppState } from '../../lib/app_model_context'
 import { useLocaleContext } from '../../lib/locale_strings'
 import { isSelfCustodyProvider } from '../../../shared/lib/external_wallet'
-import { Modal } from '../modal'
+import { Modal } from '../common/modal'
 import { PaymentForm } from './payment_form'
 import { PaymentSelection } from './payment_selection'
 import { Sending } from './sending'
