+++ components/brave_rewards/resources/rewards_page/components/modals/reset_modal.style.ts
@@ -4,26 +4,25 @@
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 import { color } from '@brave/leo/tokens/css/variables'
-import { scoped } from '../lib/scoped_css'
+import { scoped } from '../../lib/scoped_css'
 
 export const style = scoped.css`
   & {
+    overflow: auto;
+    max-width: 600px;
+    color: ${color.text.secondary};
+
+    a {
+      font-weight: 600;
+      text-decoration: none;
+    }
+
     @container style(--is-wide-view) {
       max-width: 600px;
     }
   }
 
-  .processing {
-    --leo-progressring-size: 38px;
-
-    padding: 32px 0 16px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 32px;
-  }
-
-  .status-icon {
+  .message-icon {
     --leo-icon-size: 40px;
 
     height: 88px;
@@ -35,14 +34,8 @@ export const style = scoped.css`
     color: ${color.systemfeedback.warningIcon};
   }
 
-  h3 {
-    margin: 32px 0 16px;
-  }
-
-  .error-text {
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
+  .text {
+    padding-bottom: 16px;
   }
 `
