+++ components/brave_rewards/resources/rewards_page/components/modals/reset_modal.tsx
@@ -7,11 +7,11 @@ import * as React from 'react'
 import Icon from '@brave/leo/react/icon'
 import Checkbox from '@brave/leo/react/checkbox'
 
-import { useLocaleContext } from '../lib/locale_strings'
-import { formatMessage } from '../../shared/lib/locale_context'
-import { NewTabLink } from '../../shared/components/new_tab_link'
-import { Modal } from './modal'
-import * as urls from '../../shared/lib/rewards_urls'
+import { useLocaleContext } from '../../lib/locale_strings'
+import { formatMessage } from '../../../shared/lib/locale_context'
+import { NewTabLink } from '../../../shared/components/new_tab_link'
+import { Modal } from '../common/modal'
+import * as urls from '../../../shared/lib/rewards_urls'
 
 import { style } from './reset_modal.style'
