+++ components/brave_rewards/resources/rewards_page/components/onboarding/onboarding_error_modal.tsx
@@ -10,7 +10,7 @@ import Button from '@brave/leo/react/button'
 import { EnableRewardsResult } from '../../lib/app_state'
 import { useLocaleContext } from '../../lib/locale_strings'
 import { formatMessage } from '../../../shared/lib/locale_context'
-import { Modal } from '../modal'
+import { Modal } from '../common/modal'
 import { NewTabLink } from '../../../shared/components/new_tab_link'
 import * as urls from '../../../shared/lib/rewards_urls'
