+++ components/brave_rewards/resources/rewards_page/components/onboarding/onboarding_success_modal.tsx
@@ -9,7 +9,7 @@ import Button from '@brave/leo/react/button'
 import { AppModelContext } from '../../lib/app_model_context'
 import { NewTabLink } from '../../../shared/components/new_tab_link'
 import { useLocaleContext } from '../../lib/locale_strings'
-import { Modal } from '../modal'
+import { Modal } from '../common/modal'
 import { WdpOptInModal } from './wdp_opt_in_modal'
 import * as urls from '../../../shared/lib/rewards_urls'
